Removed unnecessary Code. Moved PhalanxInfo class to utils.
