+++ package.json
@@ -36,7 +36,7 @@
     "sqlite3": "^5.0.8"
   },
   "devDependencies": {
-    "@types/node": "^18.14.1",
+    "@types/node": "^18.14.2",
     "apidoc": "^0.51.1",
     "html-loader": "^3.1.0",
     "nodemon": "^2.0.16",
