+++ userscript/ui/tabStatus.js
@@ -1,99 +1,6 @@
 const statusHtml = require('./tabStatus.html').default
 const req = require('../requests')
-const {
-  getCurrentPosition
-} = require('../utils')
-
-const ACTIVITY_INACTIVE = 0
-const ACTIVITY_BANNED = 1
-const ACTIVITY_VACATION = 2
-const ACTIVITY_ACTIVE = 3
-
-class GalaxyPhalanxInformation {
-  constructor (phalanxes) {
-    if (!Array.isArray(phalanxes)) {
-      throw new Error('Expected an array of phalanx information.')
-    }
-    this.phalanxes = phalanxes
-    this.data = Array.from(Array(400), () => [])
-    this.calculateSystemsInPhalanx()
-  }
-
-  getSystem (system) {
-    return this.data[system - 1]
-  }
-
-  appendPhalanx (system, phalanx) {
-    if (system < 1 || system > 400) return
-    return this.getSystem(system).push(phalanx)
-  }
-
-  calculateSystemsInPhalanx () {
-    this.phalanxes.forEach(phalanx => {
-      const { from, to } = phalanx.range
-      if (from < to || from === to) {
-        for (let system = from; system <= to; system++) {
-          this.appendPhalanx(system, phalanx)
-        }
-      } else {
-        for (let system = from; system <= (400 + to); system++) {
-          this.appendPhalanx(system % 401, phalanx)
-        }
-      }
-    })
-  }
-
-  isInRange (system) {
-    if (system < 1 || system > 400) return false
-    return this.getSystem(system).length > 0
-  }
-
-  getActivity (system) {
-    if (system < 1 || system > 400) return 3
-
-    const phalanxInSystem = this.getSystem(system)
-    let activity = 0
-    phalanxInSystem.forEach(phalanx => {
-      if (phalanx.galaxy === 4 && phalanx.system === 30) console.log({ phalanx, activity })
-      if (phalanx.isVacation && activity < ACTIVITY_VACATION) activity = ACTIVITY_VACATION
-      if (phalanx.isBanned && activity < ACTIVITY_BANNED) activity = ACTIVITY_BANNED
-      if (phalanx.isInactive > 0) activity = ACTIVITY_INACTIVE
-      if (!phalanx.isVacation && !phalanx.isBanned && phalanx.isInactive === 0) activity = ACTIVITY_ACTIVE
-    })
-
-    return activity
-  }
-
-  getActivityColor (system) {
-    if (system < 1 || system > 400) return 'red'
-
-    const activity = this.getActivity(system)
-    console.log(activity)
-    switch (activity) {
-      case ACTIVITY_ACTIVE:
-        return 'red'
-      case ACTIVITY_INACTIVE:
-        return '#999'
-      case ACTIVITY_VACATION:
-      case ACTIVITY_BANNED:
-        return '#659ec7'
-    }
-
-    return 'red'
-  }
-
-  // TODO: Tooltip not working yet
-  createDataTooltipContent (system) {
-    const systemData = this.getSystem(system)
-
-    let content = '<ul>'
-    systemData.forEach(phalanx => {
-      content += `<li>[${phalanx.galaxy}:${phalanx.system}:${phalanx.position}]</li>`
-    })
-    content += '</ul>'
-    return content
-  }
-}
+const { getCurrentPosition, PhalanxInfo } = require('../utils')
 
 function fetchAndDisplay (galaxy) {
   return req.searchPlanets({
@@ -107,7 +14,6 @@ function fetchAndDisplay (galaxy) {
       addRows(res)
     })
 }
-
 function insertHtml (anchor) {
   anchor.insertAdjacentHTML('beforeend', statusHtml)
 
@@ -137,7 +43,7 @@ async function addRows (planets) {
   const galaxy = document.querySelector('select#galaxy').value
 
   const phalanxes = await req.getPhalanxes(galaxy)
-  const galaxyPhalanxInfo = new GalaxyPhalanxInformation(phalanxes)
+  const galaxyPhalanxInfo = new PhalanxInfo(phalanxes)
 
   const ROWS_HEADER = 1
   let anchor = document.querySelector('table#galaxy-status').querySelectorAll('tr')[ROWS_HEADER - 1]
@@ -154,7 +60,7 @@ async function addRows (planets) {
       else if (planetCount >= 3) cls = 'color-green'
       if (oldestAge > 24 * 5) cls = 'color-orange'
 
-      const inPhalanxStyle = galaxyPhalanxInfo.isInRange(system) ? `style="border: 2px solid ${galaxyPhalanxInfo.getActivityColor(system)};"` : ''
+      const inPhalanxStyle = galaxyPhalanxInfo.isInRange(system) ? `style="border: 2px solid ${galaxyPhalanxInfo.getSystemActivityColor(system)};"` : ''
       const phalanxTooltipConent = galaxyPhalanxInfo.isInRange(system) ? `data-tooltip-content="${galaxyPhalanxInfo.createDataTooltipContent(system)}"` : ''
 
       html += `<td ${inPhalanxStyle} ><a ${phalanxTooltipConent} href="${window.location.pathname}?page=galaxy&galaxy=${galaxy}&system=${system}" class="${cls}">${system}</a></td>`
